FromUTF8
V8 の文字列の日本語化は FromUTF8 でいけた

+++ V8/StartupData.hpp
@@ -20,7 +20,7 @@ namespace JS
 		{
 
 
-			this->data.raw_size = this->buffer.size();
+			this->data.raw_size = static_cast<int>(this->buffer.size());
 			this->data.data = reinterpret_cast<const char *>(this->buffer.data());
